feat(Staking): refactor
