+++ src/component-library/Field/Field.style.tsx
@@ -1,12 +1,18 @@
 import styled from 'styled-components';
 
-import { Flex } from '../Flex';
 import { theme } from '../theme';
+import { Spacing } from '../utils/prop-types';
 
-const Wrapper = styled(Flex)`
+type StyledFieldProps = {
+  $maxWidth?: Spacing;
+};
+
+const StyledField = styled.div<StyledFieldProps>`
   position: relative;
   color: ${theme.colors.textPrimary};
   box-sizing: border-box;
+  display: inline-flex;
+  max-width: ${({ $maxWidth }) => $maxWidth && theme.spacing[$maxWidth]};
 `;
 
-export { Wrapper };
+export { StyledField };
