+++ src/pages/Staking copy/ClaimRewardsButton/index.tsx
