+++ src/pages/Staking copy/InformationUI/index.tsx
