+++ src/pages/Staking/Staking.style.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+import { Flex } from '@/component-library';
+
+const StyledWrapper = styled(Flex)`
+  max-width: 480px;
+  width: 100%;
+  margin: 0 auto;
+`;
+
+export { StyledWrapper };
