+++ src/pages/Staking/components/StakingForm/index.tsx
@@ -1,0 +1,2 @@
+export type { SwapFormProps } from './StakingForm';
+export { SwapForm } from './StakingForm';
