+++ src/pages/Staking/components/StakingLockTimeInput/index.tsx
@@ -1,0 +1,2 @@
+export type { StakingLockTimeInputProps } from './StakingLockTimeInput';
+export { StakingLockTimeInput } from './StakingLockTimeInput';
