+++ src/pages/Staking/components/index.tsx
@@ -1,0 +1,5 @@
+import { SwapForm, SwapFormProps } from './SwapForm';
+import { SwapLiquidity, SwapLiquidityProps } from './SwapLiquidity';
+
+export { SwapForm, SwapLiquidity };
+export type { SwapFormProps, SwapLiquidityProps };
